made the comment feature
made a new collection in the firebase called Comments and gave attributes of ID, Author, Date and Body. the ID will match that of the blog Id on which it was commented on. that is how I link the comment to the blog.
made a Comment component to handle comments of a specific blog.